+++ src/constants/enums.js
@@ -1,0 +1,5 @@
+export const Display = {
+  FRONTEND: 'frontend',
+  BACKEND: 'backend',
+  DB: 'db',
+};
