--- src/index.css
@@ -1,28 +1,0 @@
-.Counter {
-  display: inline-block;
-  font-family: sans-serif;
-  padding: 30px 60px;
-  margin: 15px;
-  text-align: center;
-}
-
-.Counter .count {
-  font-size: 120px;
-}
-
-.Counter .increment {
-  background: linear-gradient(#fff, #eee);
-  border-color: #ccc;
-  border-radius: 10px;
-  border-style: solid;
-  border-width: 1px;
-  color: #666;
-  font-size: 30px;
-  font-weight: bold;
-  margin-top: 15px;
-  padding: 15px;
-}
-
-.Counter .increment:active {
-  background: linear-gradient(#ddd, #eee);
-}
